chore(router): Refactor: Move analyzeRoutes to separate file
